Add files via upload
https://linuxdevices.org/free-linux-distro-targets-smartphones/ 

https://en.wikipedia.org/wiki/Rxvt

https://www.linux.com/news/four-alternative-linux-window-managers/

"FVWM

Rob Nation started FVWM way back in 1993 as a way to get a reliable windowing system on his 486 with only 4MB of RAM. He also needed to be able to display ultra-wide spectrograms for his acoustic analysis. This homebrew window manager, that ran under X, filled the bill. The project switched hands a few times as developers came and went, but it has been well-maintained.

I’ve been using FVWM2 regularly for a couple of years. Here is the scoop:

It’s fast to start, 15 seconds average on an old Pentium.
It is simple and plain.
It has multiple desktops.
It has a graphical menu bar at the top and one-touch right-click main menu.
Configuration is via a text file call .fvwm2rc. You can save a desktop setup to the file and then edit it to change the configuration.
Today’s current stable version is 2.4.18. FVWM is pretty well documented both via online how-tos and man pages.

The desktop is very mature and straightforward. If you are used to a conventional layout like KDE, the FVWM layout will be a bit of a change for you. Normally, it will start up with a “pager,” xeyes, and some push buttons on the top and a window that logs system messages. The default background is a nondescript grey stone pattern. Take a look at the Web site for cool screenshots of various configurations.

FVWM is probably the second-fastest-loading program of the group, right behind IceWM."

https://en.wikipedia.org/wiki/Hildon

https://en.wikipedia.org/wiki/Window_Maker

https://www.linux.com/news/four-alternative-linux-window-managers/

https://www.fvwm.org/

https://github.com/fvwmorg/fvwm-0.91
https://www.fvwm.org/Wiki/FvwmHistory/
"Fvwm Timeline
Fvwm was created by Robert Nation as a fork of TWM and the first version 0.5 was bundled with Rxvt and released June 1st, 1993. This version was maintained until Fvwm-0.91.
In the fall of 1993 Fvwm became an independent package and Fvwm-1.0 was released. This branch ended with the final release of Fvwm-1.24r.
In 1995 development of Fvwm2 was started and about that time Charles (Chuck) Hines took over and maintained Fvwm2 during the 2.0.x pre-release.
Fvwm switches to a version system that 2.even are stable releases while 2.odd are development releases. Starting with the development version 2.1.beta, Fvwm uses CVS to track the releases, and all releases from Fvwm 2.1.beta onward can be seen in the Fvwm2 release history.
Between spring and fall 1998 Charles Hines passes maintenance of Fvwm to Brady Montz who then passes it to Dominik Vogt.
Dominik Vogt transitions Fvwm from a single maintainer to a community model, and through a group of Fvwm-workers releases Fvwm 2.2.x in February of 1999. With the release of 2.2.x, the development version 2.3.x starts.
At this time the list of Fvwm authors gets quite large and there is no longer a single maintainer. From this point onward the Fvwm-workers community has continued to support and develop Fvwm.
On February 25, 2001, Version 2.2.5 is released. This is the last version of Fvwm with its original license before it switched to GPL-2 in version 2.4.x. Due to this license change, OpenBSD still ships 2.2.5 in its base operating system to keep Fvwm with its original license.
Version 2.4.x is released July 3, 2001 and the new development branch 2.5.x is started. This releases changes the license to GPL-2. The final release of this branch is 2.4.20 on December 9th, 2006.
Development of Fvwm slows to a crawl but the development branch 2.5.x slowly gets bug fixes and new features. Eventually this branch gets turned into the next stable release.
On August 15, 2011, version 2.6.x is released. During this release the development branch is deprecated and no longer used. Fvwm development slows to mostly bug fixes, but a few features are added to 2.6.x.
In March 2016 Fvwm moves from CVS to GitHub, where the sources currently live.
In early 2017 Fvwm 2.6.x version freezes and is put in maintenance only development. Fvwm3 is created as a fork of Fvwm2 for any future development with the intent to break compatibility with Fvwm2 to allow major changes.
In 2019 Thomas Adam has started work on Fvwm 3.x, the next major Fvwm version, which is currently in the alpha stage.|